backend/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.2.4
scikit-learn==1.6.1
statsmodels==0.14.4
yfinance==0.2.54
requests==2.32.3
reportlab==4.2.5
